code refactoring

+++ app/recipes/domain/use_case/delete_recipe.py
@@ -1,4 +1,3 @@
-from recipes.domain.exceptions.data_not_found import DataNotFound
 from recipes.domain.exceptions.recipe_not_found import RecipeNotFound
 from recipes.domain.interface.recipe_service import RecipeService
 
@@ -8,7 +7,8 @@ class DeleteRecipe:
         self.recipe_service = recipe_service
 
     def execute(self, recipe_id: int) -> None:
-        try:
-            self.recipe_service.delete(recipe_id=recipe_id)
-        except DataNotFound as e:
-            raise RecipeNotFound(f'Recipe with id -> {id} not found!')
+        if not self.recipe_service.exists(recipe_id=recipe_id):
+            raise RecipeNotFound(recipe_id=recipe_id)
+
+        self.recipe_service.delete(recipe_id=recipe_id)
+
